+++ crates/proto/src/lib.rs
@@ -63,10 +63,10 @@ pub use crate::xfer::dns_handle::{BasicDnsHandle, DnsHandle, DnsStreamHandle, St
 #[doc(hidden)]
 pub use crate::xfer::dns_multiplexer::DnsMultiplexer;
 #[doc(hidden)]
-pub use crate::xfer::retry_dns_handle::RetryDnsHandle;
-#[doc(hidden)]
 #[cfg(feature = "dnssec")]
-pub use crate::xfer::secure_dns_handle::SecureDnsHandle;
+pub use crate::xfer::dnssec_dns_handle::DnssecDnsHandle;
+#[doc(hidden)]
+pub use crate::xfer::retry_dns_handle::RetryDnsHandle;
 #[doc(hidden)]
 pub use crate::xfer::{BufDnsStreamHandle, BufStreamHandle};